hosted/stlinkv2: Began the process of splitting the JTAG and SWD specifics of the ST-Link v2 backend out into their own files
